Corregir errores de tipo de dato

+++ Practicas/Java Basic/src/imaster/clase03/sistema_notas/src/main/java/co/edu/uninorte/misiontic2022/c2/Nota.java
@@ -99,7 +99,7 @@ public class Nota {
         return escala10;
     }
 
-    public double getEscala5() {
+    public float getEscala5() {
         return escala5;
     }
 
@@ -116,7 +116,7 @@ public class Nota {
         this.escala100 = escala100;
     }
 
-    public void setEscala5(double escala5) {
+    public void setEscala5(Float escala5) {
         this.escala5 = escala5;
     }
